Calculate Beans fiels with basic features

+++ src/test/java/com/ismail/mathreflection/utilities/BeanTest.java
@@ -6,14 +6,14 @@ import com.ismail.mathreflection.annotations.Variable;
 public class BeanTest {
 
     @Variable("f1")
-    public int field1;
+    public double field1;
 
     @Variable("f2")
-    public int field2;
+    public double field2;
 
     @MXFormula("f1 - f2")
-    public int field3;
+    public double field3;
 
     @MXFormula("f1 + f2 + field3")
-    public int field4;
+    public double field4;
 }
